requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
crayons==0.4.0
cycler==0.12.1
DateTime==5.5
efficientnet_pytorch==0.7.1
facenet-pytorch==2.6.0
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
logging==0.4.9.6
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
mss==10.0.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==10.2.0
PyGetWindow==0.0.9
pyparsing==3.2.3
PyRect==0.2.0
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
toplevel==0.2.1
torch==2.2.0+cu118
torchaudio==2.2.0+cu118
torchvision==0.17.0+cu118
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
zope.interface==7.2

appdirs==1.4.4
pyinstaller==6.13.0